requirements.txt
alembic==1.14.0
Flask>=3.1.1
Flask_Bcrypt==1.0.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
google_api_python_client==2.159.0
numpy==1.26.4
passlib==1.7.4
protobuf==5.29.5
pydantic==2.10.6
PyJWT==2.10.1
python-dotenv==1.0.1
Requests>=2.32.4
SQLAlchemy==2.0.36
tensorflow==2.18.0
Werkzeug==3.1.3
email_validator==2.2.0
Flask-Migrate==4.0.7
